requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pytest
httpx
pandas
python-dotenv
pytest-cov
pillow
playwright
kiwipiepy
